+++ lib_custom/custom/index.js
@@ -1,0 +1,7 @@
+const SingleSelect = require('./SingleSelect');
+const MultiSelect = require('./MultiSelect');
+
+module.exports = {
+  SingleSelect: SingleSelect['default'],
+  MultiSelect: MultiSelect['default']
+};