+++ src/UIForm.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { connect, Form } from 'formik';
 import { clone } from './Utils';
-import * as _clone from 'lodash.clone'
+import * as _clone from 'lodash.clone';
 
 const getClasses = (use: string) => {
   const defaults = {
@@ -11,13 +11,13 @@ const getClasses = (use: string) => {
     error: 'left25pct'
   };
   if (use === 'bootstrap3' || use === 'bootstrap4') {
-    defaults.label = 'col-sm-3'
-    defaults.control = 'col-sm-9'
+    defaults.label = 'col-sm-3';
+    defaults.control = 'col-sm-9';
   }
   if (use === 'spectre') {
-    defaults.form = 'form-horizontal'
-    defaults.label = 'col-3 col-sm-12'
-    defaults.control = 'col-9 col-sm-12'
+    defaults.form = 'form-horizontal';
+    defaults.label = 'col-3 col-sm-12';
+    defaults.control = 'col-9 col-sm-12';
   }
   if (use === 'semanticui2') {
     defaults.form = 'ui form';
@@ -26,34 +26,34 @@ const getClasses = (use: string) => {
 };
 
 interface IUIFormProps {
-  use?: '' | 'bootstrap3' | 'bootstrap4' | 'spectre' | 'semanticui2',
-  horizontal?: string|boolean,
-  css?: any, // TODO: define interface for "css"
-  className?: string,
-  disabled?: string|boolean,
-  children?: any,
+  use?: '' | 'bootstrap3' | 'bootstrap4' | 'spectre' | 'semanticui2';
+  horizontal?: string | boolean;
+  css?: any; // TODO: define interface for "css"
+  className?: string;
+  disabled?: string | boolean;
+  children?: any;
   formik?: any; // FormikContext<{}>
 }
 
 const UIForm = (props: IUIFormProps) => {
   const classes = getClasses(props.use);
-  const horizontal = typeof props.horizontal !== 'undefined' ? props.horizontal.toString() : ''
+  const horizontal = typeof props.horizontal !== 'undefined' ? props.horizontal.toString() : '';
 
   props.formik.ezUse = props.use; // bootstrap, spectre, etc.
-  props.formik.ezHorizontal = horizontal
-  
+  props.formik.ezHorizontal = horizontal;
+
   const customCss = clone(props.css);
   if (props.formik.ezHorizontal) {
-    customCss.form = customCss.form || classes.form
-    customCss.label = customCss.label || classes.label
-    customCss.control = customCss.control || classes.control
+    customCss.form = customCss.form || classes.form;
+    customCss.label = customCss.label || classes.label;
+    customCss.control = customCss.control || classes.control;
   }
   props.formik.ezCss = customCss;
 
-  const clonedProps = _clone(props)
-  clonedProps.horizontal = horizontal
-  
-  const className = props.className ? `${classes.form} ${props.className}` : classes.form
+  const clonedProps = _clone(props);
+  clonedProps.horizontal = horizontal;
+
+  const className = props.className ? `${classes.form} ${props.className}` : classes.form;
   return <Form {...clonedProps} className={className} />;
 };
 export default connect(UIForm);