+++ src/examples/LayoutExample.tsx
@@ -1,18 +1,17 @@
 import * as React from 'react';
 import * as Yup from 'yup';
-import { DisplayFormState } from '../Utils'
+import { DisplayFormState } from '../Utils';
 
-import FormContainer from '../UIFormContainer'
+import FormContainer from '../UIFormContainer';
 import Form from '../UIForm';
 import Field from '../UIField';
 import Button from '../UIButton';
 
 const schema = Yup.object().shape({
-  username: Yup.string().required('Username is required!'),
+  username: Yup.string().required('Username is required!')
 });
 
 export default class extends React.Component {
-
   renderForm = (props: any) => {
     return (
       <Form use="bootstrap4">
@@ -24,12 +23,12 @@ export default class extends React.Component {
           <Field password name="password" />
           <Field password name="confirm" />
         </Field>
-  
+
         <Button type="submit" />
         <DisplayFormState {...props} />
       </Form>
-    )
-  }
+    );
+  };
 
   render() {
     return (